Clickhouse table change data type
Change the type of execution in test runs to float

+++ helm/data-plane/files/clickhouse/init-db.sql
@@ -99,7 +99,7 @@ CREATE TABLE IF NOT EXISTS test_runs (
     failCriteria String,
     testCase String, -- JSON stored as String
     score String,    -- JSON stored as String
-    executionTimeSeconds Int32 NULL,
+    executionTimeSeconds Float32 NULL,
     runAt DateTime DEFAULT now(),
     sign Int8,
     PRIMARY KEY (id)
